Refactor Subscribe logic into src/Subscriber.
